Move sql columns and tables names from Model classes to DAO classes
